Move load_costumes() to lib/servant.py
Move Costume to lib/types.py as CostumeData
