+++ lib/__init__.py
@@ -6,6 +6,7 @@ from .io import load_json, save_json
 from .merge import merge
 from .servant import (
     ServantLogger,
+    load_costumes,
     load_servants,
     servant_dict,
     skill_dict,
@@ -16,6 +17,7 @@ from .text import load_item_dictionary
 from .types import (
     AppendSkills,
     Costume,
+    CostumeData,
     CostumeID,
     Dictionary,
     Item,
